New NoteUtil is here
