--- noteutil/__init__.py
@@ -1,9 +1,0 @@
-from .errors import *
-from .notes import *
-from .extensions import *
-from .groups import *
-from .categories import *
-from .quiz import *
-
-
-name = "noteutil"
