+++ noteutil/testing/edit_test.py
@@ -1,6 +1,5 @@
-from noteutil2.noteutil import NoteUtil
-from noteutil2.comparisons import CompareOptions
-from noteutil2.errors import NoteError
+from noteutil.noteutil import NoteUtil
+from noteutil.comparisons import CompareOptions
 import os
 
 
@@ -21,11 +20,7 @@ def test_noteutil(noteutil):
     string += "----------------" + "\n"
     for pair in noteutil.pairs:
         string += "\t" + pair.term + "\n"
-    string += "--------------------------" + "\n"
-    string += "Extension Names and Bounds" + "\n"
-    string += "--------------------------" + "\n"
-    for name, bounds in zip(noteutil.extension_names, noteutil.extension_bounds):
-        string += "\t" + name + ": " + bounds[0] + " " + bounds[1] + "\n"
+    string += "----------------" + "\n"
 
     return string
 
@@ -36,7 +31,7 @@ def test_note(note):
     string += "Note Index: {!s:<5}\t\t".format(note.nindex)
     string += "Term: {!s:<10}\t\t".format(note.term[:10] if note.term else None)
     string += "Definition: {!s:<10}\t\t".format(note.definition[:10] if note.definition else None)
-    string += "Separator: {0}\t\t".format(note.separator) + "\n"
+    string += "Separator: {!s:<5}\t\t".format(note.separator)
     if note.has_extensions():
         string += "\t----------" + "\n"
         string += "\tExtensions" + "\n"
@@ -48,7 +43,7 @@ def test_note(note):
 
 
 def test_extension(ext):
-    string = "Extension:\t"
+    string = "Extension: \t"
     string += "Content: {0}\t\t".format(ext.content)
     string += "Name: {!s:<10}\t\t".format(ext.name[:10])
     string += "Left Bound: {!s:<10}\t\t".format(ext.lbound[:10])
@@ -61,7 +56,30 @@ def test_note_list(note_list):
     return "\t\n".join((list(map(test_note, note_list))))
 
 
-if os.path.exists("extensions_notes.nu"):
-    os.remove("extensions_notes.nu")
-noteutil = NoteUtil("extensions_config.txt")
+def before_after_edit(note, content):
+    print("Before:\n"
+          "-------\n" + test_note(note) + "\n" +
+          "-------")
+    noteutil.edit(note, content)
+    print("After:\n"
+          "------\n" + test_note(note) + "\n" +
+          "------")
+
+
+if os.path.exists("edit_notes.nu"):
+    os.remove("edit_notes.nu")
+noteutil = NoteUtil("edit_config.txt")
 print(test_noteutil(noteutil))
+
+n1 = noteutil.get(nindex=0)
+before_after_edit(n1, "Wumbo")
+n2 = noteutil.get(nindex=1)
+before_after_edit(n2, "Eung Eung %% APink %%")
+
+print(test_noteutil(noteutil))
+
+noteutil.reformat()
+
+print(noteutil.get(extension_names="Cool beans", compare=CompareOptions.IN))
+
+
