--- noteutil/testing/notes_testing.py
@@ -1,9 +1,0 @@
-from noteutil.noteutil.notes import Note, Line, Pair
-
-
-with open("test1,txt", mode="r") as f:
-    contents = f.read().split("\n")
-
-notes = []
-for i, content in enumerate(contents):
-    pass
