--- noteutil/testing/test1.txt
@@ -1,18 +1,0 @@
-TEST1: TEST2
-#Comment
-        # spaced comment
-
-~ Category 1 Test
-Data1: 2
-Data2: 5%93%
-! Data3: 9
-Data4: 10%01%
-
-~~ Category 2 In category 1
-Test1: 1
-! Test2: 4%99%
-Test3 : 6
-
-~ Category 1, second
-DAB: ANGLE
-some: dumb
